Rename Visual.MenuTree to ContextMenu and add Template.Menu
Visual.ContextMenu is a place to add contextual actions to the given visual: Primarily parameterless Methods on referenced objects, and Methods with a single, matching parameter from both the current and the referenced object).

Template.Menu is the "global" menu for a given Template. Template and data specific functionality can go here.


git-svn-id: svn+ssh://dasz.at/srv/dasz/svn/Kistl@403 76cbe2a5-d633-47ad-8992-a0d0e66ffbf2