Multiple changes to fix Case 1149 "After submitting changes 1:n Lists are not refreshed" and some performance issues.
On SetObjects(), the server needs to send back all changed objects. To improve the situation, the client pushes a list of items it wants to be notified about (see ObjectNotificationRequest).

Eager load only when the navigator REALLY wants to eagerly load, not when ANY navigator of the Relation specifies it.


git-svn-id: svn+ssh://dasz.at/srv/dasz/svn/Kistl@1290 76cbe2a5-d633-47ad-8992-a0d0e66ffbf2