[API,Tests,Configuration] Move all configurations to a separate area in the repository and adpat all scripts to use this
The recommended way to use this is to copy the required files from the Examples tree to "Configs\Local" and set zenv=Local in the environment. These files can be adapted to the local development environment. The CCnet* directories are used by the continuous integrations servers running in our office.

git-svn-id: svn+ssh://dasz.at/srv/dasz/svn/Kistl@2828 76cbe2a5-d633-47ad-8992-a0d0e66ffbf2

Removed customer's generated code and schema