API: Update MagicCollectionFactory to consider co-/contravariance, part 2
A List<string> is assignable to IEnumerable<object> but that's not what
we're looking for here.

Not changing this would lead to not-so-subtle errors when the MCF would
return a SynchronizedReadOnlyCollection<object> instead of a casting
wrapper when being passed a List<string> and a List<object> is requested.

This supercedes the botched last commit.