ZbFutureTask: await the future
Without this Wait(), the intermediateTask has noone waiting on its result.
This would lead him to Post his result actions, which would not be able
to run, since the Future task is already blocking the calling thread.

With this Wait(), the intermediate task can Pulse the waiter and is run
properly before the Future task runs.