QueryTranslator: implement forcing all lambdas to actually connect with the underlying source (type)
This is required to get rid of casts and similar things when working
without the real type on a query. For example, the following query
now works:

    entries = ctx.Internals().GetPersistenceObjectQuery(ifType).OrderBy(o => ((IExportable)o).ExportGuid);

Formerly, the translator would have died trying to convert the
IPersistenceObject o into an IExportable. Now it works, as long as ifType
is actually exportable.